requirements.txt
blobconverter==1.3.0
boto3==1.26.61
botocore==1.29.61
certifi==2022.12.7
charset-normalizer==3.0.1
depthai==2.20.2.0
depthai-sdk==1.9.2
distinctipy==1.2.2
googleapis-common-protos==1.58.0
grpclib==0.4.3
h2==4.1.0
hpack==4.0.0
hyperframe==6.0.1
idna==3.4
jmespath==1.0.1
marshmallow==3.17.0
multidict==6.0.4
numpy==1.24.1
opencv-contrib-python==4.7.0.68
packaging==23.0
Pillow==9.4.0
protobuf==4.21.12
python-dateutil==2.8.2
pytube==12.1.2
PyTurboJPEG==1.6.4
PyYAML==6.0
requests==2.28.2
s3transfer==0.6.0
six==1.16.0
typing_extensions==4.4.0
urllib3==1.26.14
viam-sdk==0.2.8
xmltodict==0.13.0
